LayoutTests/js/dom/document-write-functions-cached.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p>This page tests that the document functions write() and writeln() are cached. If the test passes, you'll see a PASS messages below.</p>
<pre id="console"></pre>
<script>
if (document.write === document.write)
    debug("PASS: document.write is cached.");
else
    debug("FAIL: document.write is not cached, but should be.");

if (document.writeln === document.writeln)
    debug("PASS: document.writeln is cached.");
else
    debug("FAIL: document.writeln is not cached, but should be.");

if (this.testRunner)
    testRunner.dumpAsText();
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
